n_vox x, y and z are now int. this refs #22

+++ src/pyMBIR_UI/advanced_settings/advanced_settings_initialization.py
@@ -56,13 +56,13 @@ class AdvancedSettingsInitialization:
             n_vox_z = np.NaN
         else:
             crop_width = session_dict['crop']['width']
-            n_vox_x_y_value = crop_width / vox_xy_to_use
+            n_vox_x_y_value = int(crop_width / vox_xy_to_use)
             n_vox_x = n_vox_x_y_value
             n_vox_y = n_vox_x_y_value
             n_vox_x_to_use = n_vox_x
             n_vox_y_to_use = n_vox_y
             crop_height = session_dict['crop']['to slice - from slice']
-            n_vox_z = (crop_height) / vox_xy_z_value
+            n_vox_z = int(crop_height / vox_xy_z_value)
 
         write_output_flag = config['write output']
